move jest config under tools
